Renew access token and re-try query if expired.
If the access token expires for some reason, we prefer to re-try the
query rather than forcing a redirect or simply throwing an exception. In
order for this to work without constantly refreshing the access token,
the framework developer needs to put a check for the change on the after
filter (as in, after the route completes). If there is change, update
the session hash. This is a spot annoying, but I suspect none of the
other oauth2 implementations are doing this correctly, because I can't
think of another way to implement this.